250441100142-Moh.NafisMuzadi-AsprakKakOkta

+++ modul-1/250441100142-Moh.NafisMuzadi2.py
@@ -1,6 +1,6 @@
-panjang = int(input ("masukkan panjang balok : "))
-lebar = int(input ("masukkan lebar balok : "))
-tinggi = int(input ("masukkan tinggi balok : "))
+panjang = float(input ("masukkan panjang balok : "))
+lebar = float(input ("masukkan lebar balok : "))
+tinggi = float(input ("masukkan tinggi balok : "))
 
 volume = panjang * lebar * tinggi
 
@@ -8,4 +8,3 @@ luas = 2 * (panjang * lebar + panjang * tinggi + lebar * tinggi )
 
 print("volume balok adalah :", volume, "cm")
 print("luas balok adalah :", luas, "cm")
-
